-Eye Glows now save to .chr files. Older .chr files will use the Default setting. -Eye Glow Menu will now properly update when loading models & character files. Lightwave Exporter: -Officially replaced the M2 Exporter with a superior version! -M2 Exporter now supports exporting Lights to the scene file. -M2 Character Exporter will now only export the Body texture as a character-specific image. The rest will export with the proper, reusable names. -M2 Attached Models will now export the proper image names & paths. -Added functionality for Camera & Bone Exporting options. DevWork: -Attached Models' TextureList will now properly list the used image files. -Removed Experimental Lightwave exporter menu option. No longer needed. -Removed old Lightwave M2 Exporter. -Reorganized the modelexport_lwo.cpp file. -Increased 64-bit compatibility.

+++ src/mpq_stormlib.cpp
@@ -252,12 +252,12 @@ bool MPQFile::isEof()
     return eof;
 }
 
-void MPQFile::seek(int offset) {
+void MPQFile::seek(ssize_t offset) {
 	pointer = offset;
 	eof = (pointer >= size);
 }
 
-void MPQFile::seekRelative(int offset)
+void MPQFile::seekRelative(ssize_t offset)
 {
 	pointer += offset;
 	eof = (pointer >= size);
